rename to db to be less specific if someday we switch
